kcloud-manage/AI_SW_IDE/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
kubernetes==28.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
websockets==12.0
apscheduler==3.10.4